Move StreamCipherAlgorithm to Experimental namespace
