+++ tests/Algorithms/ChaCha20Tests.cs
@@ -1,4 +1,5 @@
 using NSec.Cryptography;
+using NSec.Experimental;
 using Xunit;
 
 namespace NSec.tests.Algorithms
