+++ tests/Base/StreamCipherAlgorithmTests.cs
@@ -1,6 +1,7 @@
 
 using System;
 using NSec.Cryptography;
+using NSec.Experimental;
 using Xunit;
 
 namespace NSec.Tests.Base
