+++ tests/Rfc/ChaCha20Tests.cs
@@ -1,5 +1,6 @@
 using System;
 using NSec.Cryptography;
+using NSec.Experimental;
 using Xunit;
 
 namespace NSec.Tests.Rfc
